requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
pydantic==2.12.5
requests==2.32.5
python-dotenv==1.2.1
pandas==2.3.3
numpy==2.4.1
scikit-learn==1.8.0
scipy==1.17.0
matplotlib==3.10.8
joblib==1.5.3
yfinance==1.1.0
ta==0.11.0
tqdm==4.67.1
optuna>=3.0.0
langchain-openai>=0.3.0
langchain-anthropic>=0.3.0
langgraph>=0.4.0
langchain-core>=0.3.0
selenium>=4.0.0
undetected-chromedriver>=3.0.0
